packages: nginx: fix libopenssl dependency
The build system doesn't understand this compound DEPENDS condition
without it being wrapped in parentheses. menuconfig does understand it, so
libopenssl would be selected if nginx needed it, but there was no
guarantee that openssl would be built before nginx. Adding parentheses
fixes the problem. This seems to be how compound conditions are generally
handled in OpenWrt Makefiles.

Signed-off-by: Mark Mentovai <mark@moxienet.com>

git-svn-id: svn://svn.openwrt.org/openwrt@36606 3c298f89-4303-0410-b956-a3cf2f4a3e73