kernel: kmod-appletalk: fix dependencies
When CONFIG_LLC gets build as a module, also p8022.ko and psnap.ko are 
getting build as a module. kmod-appletalk depends on llc.ko and 
psnap.ko, but at least psnap.ko,was not packed. On most systems 
CONFIG_LLC will be build into the kernel so this problem does not show 
up.

This fixes the missing dependency of kmod-appletalk on psnap.ko

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>


git-svn-id: svn://svn.openwrt.org/openwrt@37673 3c298f89-4303-0410-b956-a3cf2f4a3e73