package/kernel: add spi-gpio-custom module
The package makes it possible to create custom SPI
buses via GPIO lines.

This module is based on i2c-gpio-custom and allows to
configure an SPI bus via GPIO lines and expose it to
userspace without rebuilding the kernel. This is maily
intended to connect microcontrollers and other simple
SPI devices without making an ad-hoc kernel driver.

Signed-off-by: Marco Burato <zmaster.adsl@gmail.com>
[juhosg:
  - change subject,
  - extend and reformat commit message,
  - cleanup checkpatch errors/warnings in module source]
Signed-off-by: Gabor Juhos <juhosg@openwrt.org>

git-svn-id: svn://svn.openwrt.org/openwrt@37872 3c298f89-4303-0410-b956-a3cf2f4a3e73